pyproject.toml
[tool.poetry]
name = "python-takehome-2"
version = "0.1.0"
description = ""
authors = ["Pedro <pedro@z33dd.com>"]
readme = "README.md"

[tool.poetry.scripts]
app = "app.main:app"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
supabase = "^2.5.1"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
psycopg2 = "^2.9.9"
celery = {extras = ["redis"], version = "^5.4.0"}
typer = "^0.12.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
pytest = "^8.2.2"
faker = "^25.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
